+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log
 
+## [Unreleased]
+
+### Changed
+
+- concord-server: forks now re-use the parent process' `imports`.
+
+
+
 ## [1.35.1] - 2019-10-17
 
 ### Changed